+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 #### Improvements
 - Upgraded dependencies
 - Use [ink-docstrap-template](https://www.npmjs.com/package/ink-docstrap-template) instead [ink-docstrap](https://www.npmjs.com/package/ink-docstrap) to deal with updated components, thus avoiding security warnings
+- Direct use of [jsdom](https://www.npmjs.com/package/jsdom) instead of [mock-browser](https://www.npmjs.com/package/mock-browser) (outdated) in `test/nodejs`, to avoid security warnings
 
 ### v2.1.5 (2021-11-07)
 #### Improvements